+++ src/main/java/tools/util/DatastoreHelper.java
@@ -54,7 +54,7 @@ public class DatastoreHelper {
         com.google.appengine.api.datastore.Entity entity = pq.asSingleEntity(); // Retrieve up to five posts
         String availableVideos1 = (String) entity.getProperty("availableVideos");
         String availableVideos = availableVideos1 + ",";
-        entity.setProperty("score", ((int)entity.getProperty("score")) + point);
+        entity.setProperty("score", ((long)entity.getProperty("score")) + point);
         entity.setProperty("availableVideos", availableVideos + url);
         try {
             datastore.put(entity); // store the entity
